src/securia_llm/requirements.txt
envyaml
requests
pyOpenSSL
json2html
prometheus_client
schedule
fastapi
cryptography
uvicorn
pydantic
pyjwt
# langchain
openai
s3fs